Bugfix: if self.end_time was None, self.end_time._to_dict() was called
That caused an AttributeError, as None has no _to_dict() method.
That happened in the deserialization process, where a single Time
(eg: 15h30) was processed. In this case, the serialied version
indicates 'end_time': None, which is passed to the TimeInterval
__init__ method.

To fix this behavior, we test if self.end_time is None or if self
has an attribute 'end_time'. In both these cases, end_time is set to
None, else, end_time = self.end_time._to_dict()
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>