Bugfix: a deserialized TimeInterval with no end_time can now be re-serialized()
The problem occured when a deserialized TimeInterval with no end_time was
re-serialized(). As it has no end_time *attribute*, the test
``if self.end_time`` caused an AttributeError.

The test is now ``if hasattr(self, 'end_time') and self.end_time``.

Deserialization unit-tests have been completed, to test this case:
we test that original.serialize() == deserialized.serialize()

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>