New feature: is a Timepoint in the future?
Each Timepoint holds a ``future`` method, returning a boolean,
indicating if it is in the past (False) or the future (True).

A ``datetime.date`` instance can be passed as ``reference``
argument, to indicate what is the time reference to be used.
The default reference is ``datetime.date.today()``, meaning the date
of the method execution.

Note that the ``future`` method only relies on the *date* and not on
the time. Thus, Time & TimeInterval instances ``future`` calls
will always return False.

Associated unit tests are sent along.
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>