Fixes #1: numeric dates with 2 digit dates now handled properly
If a numeric date has a 2 digit year format (dd/mm/yy), it is
impossible to know it the date refers to a past date or a future one.

We solve this problem this way: we consider that the date is future.
If it is located less than 15 years into the future, we leave it that
way, otherwhise, we consider it is from the previous century.

Examples:
 * 01/01/16: 1st january 2016
 * 01/01/78: 1st january 1978

A unit test has been added.

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>