New utility function is_future: indicates if a start/end interval is in the future
The is_future function returns True if the datetime interval is located in the
future (eg if the starting timepoint is located after the time reference or if
the reference is included in the start, end interval).
Else, it returns False.

Unit tests are shipped.

The function is accessible through datection.is_future for more commodity
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>