Bugfix: Fixed Redundancy in independant contexts
The list of context treated by the ``context.independants`` function
was not sorted. Thus, some redundancy could be found in its result.

Ex: [A, A, A, B, A] → [A, B, A]
where all A overlap (same goes for B).

By sorting the list using the context start position, we ensure the
independance of the result contexts:
[A, A, A, A, B] → [A, B]

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>