If year is not detected, set it to datetime.MINYEAR instead of None
In the case of a normalisation where the year is not detected, we used
to set the year to None. A ValueError was raised during the conversion
to python datetime, as a datetime does not support None year.

We resolve this issue by setting the year to datetime.MINYEAR (constant
with value 1). The Timepoint object can thus be converted to python
datetime, with year=1.

We thus test the value of year to assess the Timepoint validity, instead
of just testing if the year was defined.

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>