Handle the case of a missing year in the normalisation of a DateList
If the year is missing, we either set its value to the current year
or the value of the last date in the list.

If the last date does not have an associated year (ex: le 5, 6 et 7
octobre'), all dates occur in the current year.
Else, all dates occur in the year of the last date (ex: le 5, 6 et 7
octobre 2007)
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>