New supported french date interval syntax: "du dd/mm/yy(yy) au dd/mm/yy(yy)"
Some control has been added on the argument 'year' passed to the
Date constructor (which only happens when a Date is instanciated
via another Timepoint, and not from a regex detection): if the
year has 4 digits, accept it, but if the year has only 2 digit,
normalise it.

Some unit tests have been added to cover this case.

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>