Sanitation + enhancement of the date formatting code
A dedicated class (``datection.display.ScheduleFormatter``) is used
for the formatting of a datection schedule.

The only external API is ``datection.display.ScheduleFormatter.display``,
which transforms the result of a ``to_mongo`` export into a human
readable text block, describing the dates and times.

The best result being always the shortest, 2 techniques are used,
and only the shortest result is selected:
* display single dates and dates intervals
* display sparse date lists

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>