Normalisation of the recurrence regex matches
The basic recurrence class is ``WeekdayRecurrence``.
A weekly recurrence bears 3 important pieces of information:
 * the start datetime
 * the end datetime
 * the recurrence rule

The ``to_python`` method returns a ``datetutil.rrule.rrule`` object,
able to generate a list of *start* datetimes.
However, the rrule is not able to generate the *end* datetime for each
day. We'll have to store a "duration" or "end time" parameter and
calculate it on the fly for each start datetime.

The ``__str__`` method returns a standard RFC string, that we can
re-parse using ``dateutil.rrule.rrulestr`` to re-instanciate an
``rrule`` object.
I think we'll store this string in database (plus duration),
instead of a shitload of start/end datetimes.

The only difference between the ``WeekdayRecurrence`` and its child
classes is the way they handle setting the weekdays:
 * ``WeekdayIntervalRecurrence._set_weekdays`` returns a range between
   the first and last weekday
 * ``AllWeekdayRecurrence._set_weekdays`` returns range(0, 7) (all days)

We added a datection.utils.isoformat_concat function that generates
a RFC compliant timestamp from a datetime object.

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>