Merge of related timepoints into a single one
If several weekday recurrences were found with the same start/end
boundaries, we merge them into a global weekday recurrence describing
them all.

Then, we apply the most restrictive date boundaries on all weekday
recurrences. It is highly improbable to have a date/time interval
plus an "all {weekday} of the year". It is highly more probable that
the text was not normalized correctly and that the weekday recurrence
only applies within the date interval.
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>