Generation of a short display of the next dates
The "short=True/False" argument of the datection.display function
can now (if short == True) lead to the generation of a super short
human readable version of the schedule, bewteen the given boundaries.

Expressions like "today", "tomorrow" are used (when necessary) to
reduce the size of the result.
Note that only the next date is fully formatted. The next occurences are
only summarized.
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>