Modification of the default "UNTIL" date
The 'until' date used to be a date, and not a datetime. When dateutil
interpreted it, it converted it into a datetime, by adding the
default time (time(0, 0, 0)).

The problem with this approach is seen when dealing with bounded
recurring events (ex: tous les jours du 5 au 15 mars 2013 de 15 à 18h).
In this example, the 'until' date would be date(2013, 3, 15), which
would prevent the last datetime (datetime(2013, 3, 15, 15, 0)) to be
generated, because it would be after the until date.

By adding time(23, 59, 59) to the until date, we prevent this
boundary problem
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>