Modification of the recurrence/not recurrence decision rules
The old decision rule was: if 'BYDAY' was found in the rrule, then
it was a recurrence rule.

However, the following sentence "tous les jours du 5 au 25 mars 2013"
would have a "BYDAY" in the rrule, once converted, but is in fact
a simple date list. It can be understood as "du 5 au 25 mars 2013".

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>