New feature: export datamined text in the "to_db" result
Every time a Timepoint is datamined from human readable text and
is normalized to a DurationRRule structure, the text match is
included in the result.

This is not useful per se, but will be used further in the process,
to be able to link datamined data to their source, and position in
the webpage.
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>