Ugly fix test
We set the test date in a couple of years, at it was intended to
test a weird situation, causing a weird bug.
As there is no DTSTART in the RRule, dateutil.rrule sets it to
the present date.

As the test date is past, the display date cannot be generated,
but this is a pure fictional error, as IRL date, the "test" date
will always be the current date...
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>