Tolerate missing year, but do not replace it by next year
THis has to be the 5th or 6th we change the missing year behaviour.
Our current take is the following: datection should be able to
tolerate and export a missing year, and will not make any attempt
to guess a replacement at first.

A *second pass* will be necessary to replace missing years by their
proper replacement.
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>